fixed the test robot to use
The tests need a fully defined robot model. That means they should have all inertia parameters set properly. The puma 560 has this.
In order to not bloat the testing time, we use just the first three links. In order  to keep the model files clean, we create the robot object in the test setup.